requirements.txt
numpy
torch==2.1.0
git+https://github.com/huggingface/transformers
pillow
ipdb
accelerate
wandb
torchmetrics
ruff
datasets
deepspeed
packaging
peft
bitsandbytes
sentencepiece
matplotlib
faker
ipdb